+++ go/cmd/sidecar/ms_communication_handlers.go
@@ -1,3 +1,19 @@
+// Package main, that implements 'sidecar' functionality
+//
+// File: ms_communication_handlers.go
+//
+// Description:
+// This file contains functions to send a message to the specified AMQ target queue. This
+// is specifically used for a microservice chain to send the message to the new destination after
+// processing on the current data steward. The function will ensure the sidecar exits
+// after sending this message
+//
+// Notes:
+// This function can perhaps be more streamlined with the generic 'send' function, clearly delining that
+// the sidecar exits afterwards.
+//
+// Author: Jorrit Stutterheim
+
 package main
 
 import (
@@ -11,6 +27,18 @@ import (
 	"google.golang.org/protobuf/types/known/emptypb"
 )
 
+// SendDataThroughAMQ sends a message to the specified AMQ target queue.
+// After sending the message, it closes the stop channel to signal the end of the program,
+// the sidecar will exit.
+//
+// Parameters:
+// - ctx: Context
+// - data: message to send in the form of a MicroserviceCommunication protobuf message.
+// - s:  serverInstance pointer, used to access the AMQ channel.
+//
+// Returns:
+// - An empty protobuf message.
+// - An error if the message could not be sent, otherwise nil.
 func SendDataThroughAMQ(ctx context.Context, data *pb.MicroserviceCommunication, s *serverInstance) (*emptypb.Empty, error) {
 	logger.Debug("Starting lib.SendDataThroughAMQ")
