+++ go/cmd/sidecar/rabbit_consume.go
@@ -1,3 +1,17 @@
+// Package main, that implements 'sidecar' functionality
+//
+// File: rabbit_consume.go
+//
+// Description:
+// This file contains the functionality to stream a incoming rabbitMQ
+// message to the main container via gRPC.
+//
+// Notes:
+// This could be up for refactoring, for Microservice Chains, we switched to not streaming the messages, but just sending them.
+// Depending on if we want to keep this streaming functionality, we might need to refactor this.
+//
+// Author: Jorrit Stutterheim
+
 package main
 
 import (
@@ -7,9 +21,21 @@ import (
 	"google.golang.org/protobuf/types/known/anypb"
 )
 
-// Handle incoming AMQ messages.
-// Stream functionality to send responses to the 'main' container. Create a SideCarMessage type by making sure the type is easily accessible
-// and package the body.
+// Handle incoming AMQ messages to stream from a sidecar to the main container.
+//
+// handleResponse handles the response received from RabbitMQ and sends it to the main container via gRPC.
+// It unmarshals the received message into the provided protobuf message, creates an `anypb.Any` message from it,
+// and constructs a `pb.SideCarMessage` with the necessary fields. If there are any trace headers in the RabbitMQ
+// message, they are added to the `pb.SideCarMessage` as well. Finally, it sends the constructed message to the
+// main container via the provided gRPC stream.
+//
+// Parameters:
+// - msg: The RabbitMQ delivery message containing the response.
+// - stream: The gRPC stream used to send the response to the main container.
+// - pbMsg: The protobuf message to unmarshal the RabbitMQ message into.
+//
+// Returns:
+// - An error if there was an error unmarshalling the message or sending it via the gRPC stream, otherwise nil.
 func (s *serverInstance) handleResponse(msg amqp.Delivery, stream pb.SideCar_ConsumeServer, pbMsg proto.Message) error {
 	proto.Reset(pbMsg)
