+++ go/cmd/sidecar/rabbit_setup_functions.go
@@ -1,3 +1,16 @@
+// Package main, that implements 'sidecar' functionality
+//
+// File: rabbit_setup_functions.go
+//
+// Description:
+// This file contains functions to establish a connection to RabbitMQ,
+// set up a topic exchange, queue, and consumer to listen to messages
+// with the specified routing key.
+//
+// Notes:
+//
+// Author: Jorrit Stutterheim
+
 package main
 
 import (
@@ -9,6 +22,18 @@ import (
 	amqp "github.com/rabbitmq/amqp091-go"
 )
 
+// getAMQConnectionString returns the AMQP connection string based on the environment variables.
+// It retrieves the AMQ_USER and AMQ_PASSWORD from the environment and constructs the connection string using the rabbitDNS and rabbitPort variables
+// that are stored in the config files.
+func getAMQConnectionString() (string, error) {
+	user := os.Getenv("AMQ_USER")
+	pw := os.Getenv("AMQ_PASSWORD")
+
+	return fmt.Sprintf("amqp://%s:%s@%s:%s/", user, pw, rabbitDNS, rabbitPort), nil
+}
+
+// getConnectionToRabbitMq establishes a connection to RabbitMQ and returns the connection and channel objects.
+// It retrieves the connection string from an environment variable, and creates the connection.
 func getConnectionToRabbitMq() (*amqp.Connection, *amqp.Channel, error) {
 	logger.Debug("Start getConnectionToRabbitMq")
 	connectionString, err := getAMQConnectionString()
@@ -20,16 +45,6 @@ func getConnectionToRabbitMq() (*amqp.Connection, *amqp.Channel, error) {
 	var conn *amqp.Connection
 	var channel *amqp.Channel
 	conn, channel, err = connect(connectionString)
-
-	// for i := 1; i <= 7; i++ { // maximum of 7 retries
-	// 	conn, channel, err = connect(connectionString)
-	// 	if err == nil {
-	// 		break // no error, break out of loop
-	// 	}
-
-	// 	logger.Sugar().Infof("Failed to connect to RabbitMQ: %v", err)
-	// 	time.Sleep(10 * time.Second) // wait for 10 seconds before retrying
-	// }
 
 	if err != nil {
 		logger.Sugar().Fatalw("Failed to setup proper connection to RabbitMQ after 7 attempts: %v", err)
@@ -80,10 +95,6 @@ func setupConnection(queueName string, routingKey string, queueAutoDelete bool) 
 			logger.Sugar().Fatalw("Queue Bind: %s", err)
 			return nil, nil, nil, err
 		}
-		// if err != nil {
-		// 	logger.Sugar().Fatalw("Queue Bind: %s", err)
-		// 	return nil, nil, nil, err
-		// }
 
 		logger.Sugar().Infof("Failed to connect to QueueBind: %v", err)
 		time.Sleep(2 * time.Second)
@@ -91,6 +102,8 @@ func setupConnection(queueName string, routingKey string, queueAutoDelete bool) 
 	return nil, conn, channel, nil
 }
 
+// connect establishes a connection to an AMQP server using the provided connection string.
+// It returns the connection and channel objects, or an error if the connection fails.
 func connect(connectionString string) (*amqp.Connection, *amqp.Channel, error) {
 	conn, err := amqp.Dial(connectionString)
 	if err != nil {
@@ -104,6 +117,8 @@ func connect(connectionString string) (*amqp.Connection, *amqp.Channel, error) {
 	return conn, channel, nil
 }
 
+// declareQueue declares a new queue with the given name, channel, and autoDelete flag.
+// It returns a pointer to the declared queue and any error encountered.
 func declareQueue(name string, channel *amqp.Channel, autoDelete bool) (*amqp.Queue, error) {
 	queue, err := channel.QueueDeclare(
 		name,       // name
@@ -121,6 +136,8 @@ func declareQueue(name string, channel *amqp.Channel, autoDelete bool) (*amqp.Qu
 	return &queue, nil
 }
 
+// exchange declares a topic exchange on the given channel.
+// It takes a pointer to an amqp.Channel and returns an error if any.
 func exchange(channel *amqp.Channel) error {
 	if err := channel.ExchangeDeclare(
 		"topic_exchange",
@@ -136,6 +153,8 @@ func exchange(channel *amqp.Channel) error {
 	return nil
 }
 
+// close_channel closes the given AMQP channel and connection after all running messages have finished.
+// It waits until there are no more running messages before closing the channel and connection.
 func close_channel(channel *amqp.Channel, conn *amqp.Connection) {
 	for {
 		if running_messages == 0 {
@@ -146,11 +165,4 @@ func close_channel(channel *amqp.Channel, conn *amqp.Connection) {
 		logger.Sugar().Infof("Waiting for %d messages to finish", running_messages)
 		time.Sleep(1 * time.Second)
 	}
-}
-
-func getAMQConnectionString() (string, error) {
-	user := os.Getenv("AMQ_USER")
-	pw := os.Getenv("AMQ_PASSWORD")
-
-	return fmt.Sprintf("amqp://%s:%s@%s:%s/", user, pw, rabbitDNS, rabbitPort), nil
 }
